fix bug for tweet gen

+++ common/src/main/java/bigframe/bigif/BigDataInputFormat.java
@@ -20,7 +20,7 @@ public class BigDataInputFormat extends Config {
 
 	protected String app_domain;
 	protected Set<String> dataVariety = new HashSet<String>();
-	protected Integer dataVolume; // In GB
+	protected Float dataVolume; // In GB
 	protected Map<String, Float> dataVelocity = new HashMap<String, Float>();
 	protected Map<String, Float> dataScaleProportions = new HashMap<String, Float>();
 	protected Map<String, String> dataStoredPath = new HashMap<String, String>();
@@ -50,7 +50,7 @@ public class BigDataInputFormat extends Config {
 		return dataVariety;
 	}
 
-	public Integer getDataVolume() {
+	public Float getDataVolume() {
 		return dataVolume;
 	}
 
@@ -224,7 +224,7 @@ public class BigDataInputFormat extends Config {
 			else if (key.equals(BigConfConstants.BIGFRAME_DATAVOLUME)) {
 //				dataVolume = BigConfConstants.DATAVOLUME_MAP.get(value);
 				
-				dataVolume = Integer.parseInt(value);
+				dataVolume = Float.parseFloat(value);
 			}
 
 			else if (BigConfConstants.BIGFRAME_DATAVOLUME_PORTION_SET.contains(key)) {
